Various changes to make gcc-7 happy.
There were a few issues that a modern gcc
found with the code, such as string truncation,
unused parameters and variables, and signed
vs unsigned comparisons, particularly for strings,
that were unneeded. Also, one possible bug in the
test code with respect to array size was
found by the compiler and fixed.