lib/sysfs_class: fix build failure reported on GCC-11
Christopher reported that, the build fails with GCC-11. It introduces a
-Wstringop-overread, that warns about reading past the end in string
functions. cdev_name_equal() does a comparison, which compares strings
past SYSFS_NAME_LEN. It currently calculates the string lengths of both
string and does a comparison on strings of equal length.

This patch refactors the function to compare, the strings until
SYSFS_NAME_LEN, removing the string length comparison logic.

Fixes #12 ("Build failure with GCC-11 due to stringop-overread")
Reported-by: Christopher Engelhard <ce@lcts.de>
Tested-by: Christopher Engelhard <ce@lcts.de>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>