build: Move test program into check_PROGRAMS
These are unit tests which should not end up installed on the system,
and should not be built as part of «make all». These will be built as
part of «make check» instead.

Ideally they would also be executed as part of the test suite, but they
currently depend on the host systems' sysfs.