build: Add support for version script
This adds a new version script so that only public symbols are exported
and no private symbols leak. The first version node provided is based on
the first version that shipped with these symbols, and new version nodes
are expected to be added for each release adding them, linking the old
one to the new one.

Signed-off-by: Guillem Jover <guillem@hadrons.org>